+++ gulpfile.js
@@ -13,6 +13,7 @@ const rename = require('gulp-rename');
 const imagemin = require('gulp-imagemin');
 const rollup = require('gulp-better-rollup');
 const sourcemaps = require('gulp-sourcemaps');
+const mocha = require('gulp-mocha');
 
 gulp.task('style', function () {
   gulp.src('sass/style.scss')
@@ -59,6 +60,12 @@ gulp.task('animate', function () {
 });
 
 gulp.task('test', function () {
+    return gulp
+        .src(['js/**/*.test.js'], { read: false })
+        .pipe(mocha({
+            compilers: ['js:babel-register'], 
+            reporter: 'spec'       
+        }));
 });
 
 gulp.task('imagemin', ['copy'], function () {
