+++ js/templates/level/artistLevelTemplate.js
@@ -1,10 +1,4 @@
 import createArtistAnswers from '../../utils/createArtistAnswers';
-import renderScreen from '../../utils/renderScreen';
-import levelType from '../../data/types/levelType';
-import createElementFromTemplate from '../../utils/createElementFromTemplate';
-import failResultTemplate from '../../templates/result/failResultTemplate';
-import headerTemplate from '../../templates/parts/headerTemplate';
-import {levels} from '../../data/data';
 
 const artistLevelTemplate = (level) => {
   return `<section class="main main--level main--level-artist">
@@ -37,28 +31,3 @@ const artistLevelTemplate = (level) => {
 
 export default artistLevelTemplate;
 
-export const artistLevelHandler = (template, state, sectionMain) => {
-  const answerButtonsWrapper = sectionMain.querySelector(`.main-list`);
-  const rightAnswer = sectionMain.querySelector(`[data-right-answer]`);
-
-  window.initializePlayer(rightAnswer, levels[state.level].rightAnswer.path);
-  answerButtonsWrapper.addEventListener(`click`, handleAwnswer.bind(answerButtonsWrapper, template, state, rightAnswer, sectionMain));
-};
-
-const handleAwnswer = function (...args) {
-  const [template, state, rightAnswer, sectionMain, event] = args;
-  const itemValue = event.target.dataset.answer;
-  if (event.target.tagName.toLowerCase() === `img`) {
-    if (itemValue === rightAnswer.dataset.rightAnswer) {
-      renderScreen(template, Object.assign({}, state, {level: levelType.Genre}));
-    } else {
-      state.lives--;
-      if (state.lives === 0) {
-        renderScreen(failResultTemplate, state);
-      } else {
-        sectionMain.replaceChild(createElementFromTemplate(headerTemplate(state)), sectionMain.firstChild);
-      }
-    }
-  }
-  this.removeEventListener(`click`, handleAwnswer.bind(this, template, state, rightAnswer, sectionMain));
-};
