+++ js/templates/start/welcomeTemplate.js
@@ -1,5 +1,4 @@
 import logoTemplate from '../parts/logoTemplate';
-import renderScreen from '../../utils/renderScreen';
 
 const welcomeTemplate = (level) => `<section class="main main--welcome">
   ${logoTemplate()}
@@ -14,13 +13,3 @@ const welcomeTemplate = (level) => `<section class="main main--welcome">
 
 export default welcomeTemplate;
 
-export const welcomeHandler = function (template, state) {
-  const playButton = document.querySelector(`.main-play`);
-  playButton.addEventListener(`click`, startGame.bind(playButton, template, state));
-};
-
-const startGame = function (...args) {
-  const [template, state] = args;
-  renderScreen(template, Object.assign({}, state, {}));
-  this.removeEventListener(`click`, startGame.bind(this, template, state));
-};
