+++ package.json
@@ -11,6 +11,8 @@
   },
   "devDependencies": {
     "autoprefixer": "6.7.7",
+    "babel-plugin-transform-es2015-modules-commonjs": "6.24.1",
+    "babel-register": "6.24.1",
     "browser-sync": "2.18.8",
     "css-mqpacker": "6.0.0",
     "del": "2.2.2",
@@ -20,6 +22,7 @@
     "gulp-better-rollup": "1.1.1",
     "gulp-csso": "3.0.0",
     "gulp-imagemin": "3.2.0",
+    "gulp-mocha": "4.3.1",
     "gulp-plumber": "1.1.0",
     "gulp-postcss": "7.0.0",
     "gulp-rename": "1.2.2",
@@ -28,5 +31,10 @@
   },
   "engines": {
     "node": "6.10"
+  },
+  "babel": {
+    "plugins": [
+      "transform-es2015-modules-commonjs"
+    ]
   }
 }
